Handle multiple documents for the same date
If you get a bonus it will typically show up the same day as your
regular pay stub. Previously you would only download one of these. Now
we keep track of how many files we have for each pay date and add
another value to the file name. If you got a bonus and a pay check on
Jan 5th 2017 you'd see `2017-01-05.pdf` and `2017-01-05-0.pdf` after
this patch.

This fixes #1.